requirements.txt
django
pandas
openpyxl
djangorestframework